+++ src/scmdata/offsets.py
@@ -15,6 +15,7 @@ def generate_range(
     start: cftime.datetime,
     end: cftime.datetime,
     offset: cftime_offsets.BaseCFTimeOffset,
+    date_cls: cftime.datetime = cftime.DatetimeGregorian,
 ) -> Iterable[cftime.datetime]:
     """
     Generate a range of datetime objects between start and end, using offset to
@@ -36,10 +37,13 @@ def generate_range(
     offset:
         Offset object for determining the timesteps.
 
+    date_cls : :class:`cftime.datetime`
+            The time points will be returned as instances of ``date_cls``
+
     Yields
     ------
     :class:`cftime.datetime`
-        Next datetime in the range
+        Next datetime in the range (the exact class is specified by ``date_cls``)
 
     Raises
     ------
@@ -62,12 +66,12 @@ def generate_range(
     ... )
 
     >>> pprint([d for d in g])
-    [cftime.datetime(2001, 1, 1, 0, 0),
-     cftime.datetime(2002, 1, 1, 0, 0),
-     cftime.datetime(2003, 1, 1, 0, 0),
-     cftime.datetime(2004, 1, 1, 0, 0),
-     cftime.datetime(2005, 1, 1, 0, 0),
-     cftime.datetime(2006, 1, 1, 0, 0)]
+    [cftime.DatetimeGregorian(2001, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2005, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2006, 1, 1, 0, 0, 0, 0)]
 
     In this example the first timestep is rolled back to 31st Dec 2000 whilst the last
     is extended to 31st Dec 2005.
@@ -78,12 +82,12 @@ def generate_range(
     ...     to_offset("A"),
     ... )
     >>> pprint([d for d in g])
-    [cftime.datetime(2000, 12, 31, 0, 0),
-     cftime.datetime(2001, 12, 31, 0, 0),
-     cftime.datetime(2002, 12, 31, 0, 0),
-     cftime.datetime(2003, 12, 31, 0, 0),
-     cftime.datetime(2004, 12, 31, 0, 0),
-     cftime.datetime(2005, 12, 31, 0, 0)]
+    [cftime.DatetimeGregorian(2000, 12, 31, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2001, 12, 31, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 12, 31, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 12, 31, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 12, 31, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2005, 12, 31, 0, 0, 0, 0)]
 
     In this example the first timestep is already on the offset so stays there, the last
     timestep is to 1st Sep 2005.
@@ -94,26 +98,27 @@ def generate_range(
     ...     to_offset("QS"),
     ... )
     >>> pprint([d for d in g])
-    [cftime.datetime(2001, 4, 1, 0, 0),
-     cftime.datetime(2001, 7, 1, 0, 0),
-     cftime.datetime(2001, 10, 1, 0, 0),
-     cftime.datetime(2002, 1, 1, 0, 0),
-     cftime.datetime(2002, 4, 1, 0, 0),
-     cftime.datetime(2002, 7, 1, 0, 0),
-     cftime.datetime(2002, 10, 1, 0, 0),
-     cftime.datetime(2003, 1, 1, 0, 0),
-     cftime.datetime(2003, 4, 1, 0, 0),
-     cftime.datetime(2003, 7, 1, 0, 0),
-     cftime.datetime(2003, 10, 1, 0, 0),
-     cftime.datetime(2004, 1, 1, 0, 0),
-     cftime.datetime(2004, 4, 1, 0, 0),
-     cftime.datetime(2004, 7, 1, 0, 0),
-     cftime.datetime(2004, 10, 1, 0, 0),
-     cftime.datetime(2005, 1, 1, 0, 0),
-     cftime.datetime(2005, 4, 1, 0, 0),
-     cftime.datetime(2005, 7, 1, 0, 0)]
+    [cftime.DatetimeGregorian(2001, 4, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2001, 7, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2001, 10, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 4, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 7, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2002, 10, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 4, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 7, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2003, 10, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 4, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 7, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2004, 10, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2005, 1, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2005, 4, 1, 0, 0, 0, 0),
+     cftime.DatetimeGregorian(2005, 7, 1, 0, 0, 0, 0)]
     """
-    # Uses the Gregorian calendar - allows for adding/subtracting datetime.timedelta in range calc
+    # Uses the Gregorian calendar - allows for adding/subtracting
+    # datetime.timedelta in range calc
     start_cf = cftime.DatetimeGregorian(*start.timetuple()[:6])
     end_cf = cftime.DatetimeGregorian(*end.timetuple()[:6])
 
@@ -121,4 +126,4 @@ def generate_range(
         offset.rollback(start_cf), offset.rollforward(end_cf), freq=offset
     )
 
-    return [cftime.datetime(*dt.timetuple()[:6]) for dt in res]
+    return [date_cls(*dt.timetuple()[:6]) for dt in res]